requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.4
python-multipart